refactor: rename smooth to scroll
