+++ lua/plugins/init.lua
@@ -56,7 +56,7 @@ require('plugins.behavior.autopairs')
 require('plugins.behavior.cmp')(which.cmp)
 require('plugins.behavior.comment')
 require('plugins.behavior.luasnip')
-require('plugins.behavior.smooth')
+require('plugins.behavior.scroll')
 require('plugins.behavior.surround')
 require('plugins.behavior.telescope')(register, which.telescope)
